+++ tests/market_stream/test_book_depth_stream.py
@@ -170,7 +170,7 @@ class TestBookDepthStream(unittest.TestCase):
 
         # Consume until logout message
         messages = client_md.retrieve_messages_until(
-            message_type="5", timeout_seconds=1
+            message_type=["5"], timeout_seconds=1
         )
 
         # Logout acknowledgment