+++ tests/market_stream/test_book_ticker_stream.py
@@ -163,7 +163,7 @@ class TestBookTickerStream(unittest.TestCase):
 
         # Consume until logout message
         messages = client_md.retrieve_messages_until(
-            message_type="5", timeout_seconds=1
+            message_type=["5"], timeout_seconds=1
         )
 
         # Logout acknowledgment